feat(build-cli): Add --exact flag to bump command
The --exact flag allows setting the version of a package or release group to a precise version. This
is useful in the CI pipeline where we need to adjust the versions for dev/test/etc. as part of the
build. This is currently done with lerna directly; this command will replace that in our build
pipeline.