feat(build-cli): Add --exact flag to bump command (#12667)
The --exact flag allows setting the version of a package or release
group to a precise version. This is useful in the CI pipeline where we
need to adjust the versions for dev/test/etc. as part of the build. This
is currently done with lerna directly:

https://github.com/microsoft/FluidFramework/blob/9080963a1194be9d02fdfec69026cb482a1cabfa/tools/pipelines/templates/include-set-package-version.yml#L92-L112

This new command will replace that in our build pipeline. We need to do
this because lerna produces invalid version ranges for Fluid internal
versions.

This change required some upstream changes in Fluid internal version
calculations. In particular, I needed to relax the requirement that the
prerelease identifier be "internal" for Fluid internal versions.
Fortunately, I was able to do this with minimal additions to existing
APIs; there is now a flag on some APIs that will allow/disallow
arbitrary prerelease identifier strings. Existing calls will continue to
enforce the "internal" prerelease identifier. To use a custom one, you
must opt in by setting the flag to true.