Forest: a workaround to notify dependents after the Delta is applied to the Forest (#12623)
This workaround enables a very simple "notification"-like mechanics to
let clients using the new SharedTree to know that the Forest has been
changed.

It is implemented with a function, which is not exported elsewhere on
purpose, as this totally misuses the invalidation system. All this code
must be removed as soon as a proper notification/eventing system will be
introduced.

Note that `afterChangeToken` is exported into the core (but not to the
tree), as it will be used by `EditableTree` to safely distinguish
"afterChange" and `beforeChange` invalidations. I'd just like to be on a
safe side in case somebody would start using invalidation tokens in the
forest.