Remove support for edit virtualization from legacy SharedTree (#12642)
- Deprecates asynchronous APIs on EditLog
- Removes edit chunking and virtualization from EditLog
- Stores all edits within a single chunk in the summary so that summary
format 0.1.1 is still supported
- Ignores edit handles when loading a summary

I did my best to make this not a breaking change when it comes to the
API but removing edit virtualization functionality means that this is a
breaking change for any clients that are using virtualization and
history. However, there shouldn't be any such clients so I made some
choices like ignoring existing edit handles to reduce complexity.