Tree: Remove modules.md (#12661)
## Description

This removes the modules.md. Its content is either already covered in
existing per module documentation, or moved into `feature-libraries`
readme.

The moved content has been edited somewhat, but the scenario list
remains is an unfinished state. It seems better to have it in its new
location that where it was, even if it's not done.