fix(build-cli): Add baseline version and normalize JSON (#12682)
Updates the typetests command to add the baseline version as a property
in the typetests node of package.json. The baseline version is helpful
information and previously you had to parse it out of the \*-previous
dependency string.

This change also always appends a trailing newline when writing JSON
files, which is consistent with our repo settings.