fix(build-cli): Add option to include .generated in type test file names (#12717)
This PR makes generated files clear from their file name so reviewers
can see the generated file warning when reviewing deltas in PRs. This
change adopts the .generated.* convention from
https://github.com/microsoft/TypeScript/wiki/Coding-guidelines#components.

I've added new flags to both the old fluid-type-validator CLI and the
new flub generate typetests entrypoint. The default for the old CLI is
false, so files will continue to have their old names by default. The
default for the new CLI is true, but it can be set to false manually
when needed while we transition to the new CLI. When we fully adopt the
new typetests tool in CI we will rename the files.