Add eslint rules to require source-code docs on *most* module-exported members (#12708)
Adds linter settings for the azure release group packages (just the
public ones), which require source-code documentation on most kinds of
module-exported members.

The tooling has some issues around variable assignments, and properties
/ methods on interfaces / classes, so there is currently no enforcement
on those. But this will ensure most root-level exports are documented at
the type-level, and will offer a friendly reminder to developers that it
is important to document the contents of our libraries (even if they
don't ultimately get exported as a part of the package).