build(client): Use updated check policy command in npm scripts (#12445)
Uses the `flub check policy` command in build scripts, replacing the
`fluid-policy-check` command.

The assert-short-codes policy is now excluded by default. It can be run
manually if needed using `npm run policy-check:asserts`. Assert tagging
should only be done in main and as part of a client release, so this
change removes it from the standard commands.