Fix the Routing Error by Setting the isSessionAlive when open a session (#12737)
When we open a session in Deli LambdaFactory, we mark both
isSessionAlive and isSessionActive as true.

There is a session in cluster 1 that ends at time t0, another session
that starts at t0+4hrs, and a third session that begins at t0+4hrs+a
couple of mins.
In this case, when the last session reads the truth, the lastAccessTime
is > 3hrs (since session 2 is not finished and lastAccessTime is updated
only on the session end). This is fine but the strange thing is that we
are seeing that when session 3 reads the truth, it has
"isSessionAlive":false, and"isSessionActive":true. This should be a
"never" case. This is weird because we thought session 2 would have made
"isSessionAlive": true after it went through discovery.
We would like to introduce this fix to mark both isSessionAlive and
isSessionActive as true, when opening a session in the deli lambda
factory so that it would still use the urls stored in the truth.


![MicrosoftTeams-image](https://user-images.githubusercontent.com/66701969/199107141-492da277-9f6f-4524-befb-0c64f0198f55.png)