feat(build-cli): Add autocomplete support for bash and zsh (#12748)
Adds an `autocomplete` command to flub which can be used to generate
autocomplete config for bash and zsh. This is provided "for free" by
using
[@oclif/plugin-autocomplete](https://github.com/oclif/plugin-autocomplete).

fish is not supported, but possibly could be with this PR:
https://github.com/oclif/plugin-autocomplete/pull/38 We could either
wait for the feature to come, release a custom plugin with that PR in
the meantime, or open our own PR to add it.