Simplify batching logic in PendingStateManager (#12592)
# [AB#1833](https://dev.azure.com/fluidframework/internal/_workitems/edit/1833)

## Description

Batch tracking inside the `PendingStateManager` will now be done through the batch metadata on ops to indicate starts and ends of batches.

Before, we were using `flush` calls to indicate a batch had ended because the `batch: false` metadata would not be updated in messages in the `PendingStateManager`. Due to changes in the batching workflow in `ContainerRuntime`, the `batch: false` metadata is available.

## Does this introduce a breaking change?

`IPendingFlush` has been deprecated. Use batch metadata on `IPendingMessage` instead to indicate the end of a batch.