Batch compression (#12159)
## Description

This is an implementation of compressing batches of messages before sending them to gain the benefits of SOLID compression using LZ4. 

The configuration aspect isn't all the way there (we have an enum for compression algorithms but in actuality are always assuming LZ4 for now). I'm erring on the side of "we can implement that when it comes up," but if we want to see some code around handling different algorithms I can include that as part of this work. 

## Reviewer Guidance

From my preliminary look, to add the top level compression property to the message it looks like we'll need a loader change to pass in the compression property when constructing the `IDocumentMessage`. Is this assumption correct or am I missing something?