Handle the treesLatestDeferral in epoch tracker (#12846)
## Description

Inside EpochTrackerWithRedemption, treesLatestDeferral.reject() is
called without handling the rejected promise. This results in an
un-intentional unhandledRejection event being triggered. This PR
resolves this issue by handing the rejected promise.

The actual response to join session/trees latest request or other calls
are truly returned. This promise was more of waiting logic in case join
session call is made before the trees latest call. This promise is imp
only in this class and its rejection does not need to be propagated. So,
we just need to add a empty catch statement so as to avoid
unhandledPromiseRejection error