Fix compression path for attachBlob ops (#12852)
## Description
The recently-merged batch compression PR caused an ICM incident because of an incorrect assumption in the batch continuation path. The continuation path checked for `metadata === undefined` rather than `metdata.batch === undefined` and ops with metadata weren't correctly detected as a batch continuation. 

I've also added an end to end test that would've caught this.