Adding GitRest metrics (part 1) (#12860)
## Description

Adding summary-related metrics for GitRest, focusing on the following
events:
- PersistLatestFullSummaryInStorage
- RetrieveLatestFullSummaryFromStorage
- WholeSummaryManagerReadSummary
- WholeSummaryManagerWriteSummary

The goal here is to collect performance data around the different
GitRest flows, including when GitRest spends time on computation and
also when communicating with storage.
This is the first iteration on this topic. I will follow up with more
pull request soon (e.g. I plan to also add optional metrics at the
`RepositoryManager` layer).