feat(fluid-build): Enable fluid-build to use pnpm and yarn (#12874)
This has been broken out of PR #12236.

This change enables fluid-build to use pnpm or yarn to install a release
group's dependencies instead of npm. pnpm will only be used if there is
a pnpm-workspace.yaml file in the root of the release group. yarn will
only be used if there is a yarn.lock file in the root of the release
group/package. Otherwise npm will be used.

This change also includes some lint related changes that were uncovered by
pnpm.