Adjusting the SummarizeFetchValidation test to expect fetch due to caching when getting versions. (#12912)
## Description

The test SummarizeFetchValidation was originally based on the local
driver run which does not make use of cache. By not using cache, as soon
as the summarizer was connected, there would be no need to fetch the
snapshot afterwards.

ODSP  does retrieve from cache, which causes
handleSummaryAck to be called and actually executes an additional fetch
when a new summarizer starts (after a previous summarizer had created a
previous summary for which we have the lastAck stored).

As the main goal of the test is to validate and control the number of
times we fetch a snapshot when running a summarizer, we want it to rely
on the services and not the local driver. For this reason, we will
bypass the local driver runs in which we would not fetch a snapshot as
it does not make use of caching.

@agarwal-navin after our conversation yesterday, I realized we do not
need to pre-process the fetch from the latest snapshot before starting
the summarizer. The reason for that is because the actual processing
starts later (depending on the heuristics or when the summarizeOnDemand
gets invoked). There might be a problem still in case we summarize
immediately after returning from the .start call - I'm still testing it
but it is out of the scope of this change.