Performance pipeline using artifacts built by previous pipelines (#12778)
This PR allows us to reuse artifacts that are created in the pipeline
that builds our client packages to run the memory and runtime
performance tests on.

There are some changes that we've had to make the repo structure in
order for this to function. I have written them down here:
**Execution Time Profiling Tests**

1. For both `experimental-tree` and `tree`: The packed `dist/test`
folder does not have the json files necessary to run the benchmark tests
so we have to change the postpack script to copy over the .ts files from
`src/test` and publish them as `test_files`, and in the pipeline, we
have to run a separate `npm build:test` to build the test files before
running the benchmark tests. See: microsoft/FluidFramework@996ba23
2. `experimental-tree` does not have a `buidl:test` command so we faked
it and introduced `build:test : npm run build` to the package.json so
that it could follow the same pattern at the others. See
microsoft/FluidFramework@f0642e790d407657bc8209f6ef5b4d00477c3679

Because of number 1 above, we didn't have to make the following changes,
but I am recording those changes here so we know about them:

**Memory Profiling Tests**

- For all the three packages (map, sequence and matrix), the packed
`dist/test` folder does not have the .mocharc.js file necessary to run
the test, so we need to move the `.mocharc.js` file from `src/test/`
files up to the root directory of the package in order to run the npm
command memory commands to work. See:
microsoft/FluidFramework@f4af0671d7cf3c95acf4aa1396981f3b970d3caa

Currently, the runs take about 24 minutes, which is longer than the
pipeline that builds everything from scratch (generally 21 minutes) as
there is no parallelization in this current approach.

Co-authored-by: Alex Villarreal <716334+alexvy86@users.noreply.github.com>