Unpack test files from `dist/test` instead of simply `test/` when unpacking test files for test pipelines (#12951)
In https://github.com/microsoft/FluidFramework/pull/12778 I changed the
structure of the test-files to pack both `src/` and `dist/` and place
them into a `test/` directory. Other pipelines expect to unpack the dist
files directly from the test directory. Since this is no longer the
case, I am updating the files that consume the test files to read them
from `test/dist/` in this PR.