feat(generate:typetests)!: Make type tests configurable per-branch (#12849)
Type tests can now be configured to use different baseline versions on a
given branch depending on the type of release that
the branch is designated for. For example, for the client release group,
the _next_ branch is the _major version series
branch_ and _main_ is the _minor version series branch_. This can be
declared in the release group root package.json, in the
`fluidBuild.branchReleaseTypes` section. For example, the following
configuration designates the _main_ and _lts_ branches as
minor version series branches, while the _next_ branch is designated for
major releases.

```json
"fluidBuild": {
  "branchReleaseTypes": {
    "main": "minor",
    "lts": "minor",
    "release/**": "patch",
    "next": "major"
  }
}
```

The type test generator takes this information into account when
calculating the baseline version to use when it's run
from a particular branch. Baseline versions are set as follows based on
the branch release designation:

| Branch release designation | Baseline version | Example: version 2.3.4
|
| -------------------------- | ---------------- |
------------------------------------ |
| patch                      | `previousPatch`  | **2.3.3**  |
| minor                      | `^previousMinor` | **^2.2.0** |
| major                      | `^previousMajor` | **^1.0.0** |

BREAKING CHANGE: `fluid-type-validator` is deprecated. Use `flub
generate typetests` instead.