fix(fluid-build): Parse build-cli tasks properly in fluid-build (#12988)
The new build-cli requires a bit more parsing to decompose the task into
fluid-build's build graph. This change updates the parsing method to
handle flub correctly.

The former implementation assumed that the command executable was the
first argument of the command, but that doesn't hold for commands with
subcommands.

The changed implementation assumes flags come at the end of the command,
and that all subsequent arguments are flags. All non-flag arguments are
then assumed to be the executable name. This is true for most of the
commands today (and all of the ones that fluid-build needs to
understand), but it is somewhat fragile.