build: policy-check should include assert tagger (#12973)
The policy-check command should, in general, include all policy
handlers. However, assert tagging is a bit unique because we only want
to run it as part of releases. An earlier change excluded the handler,
but we still need it to run in "non-fix mode" when policy-check is run,
so that it can catch issues like duplicate tags.