[TaskManager] Handle detached scenario (#12786)
## Description

This PR handles TaskManager sending ops while in a detached state. 

### Behavior while detached

If a `TaskManager` is in a detached state, then we know that the
detached client is the only one that "knows" about that `TaskManager`.
Therefore, we no longer need to send an op and wait for it to be
acknowledged by the other connected clients (since there are none).
Hence, in a detached state TaskManager will "auto-ack"/simulate outgoing
ops. This essentially means we skip submitting the op and go straight to
the logic that normally takes place after the op is ack'd.