add prettier to packages/dds (#12998)
### Description

Board:
https://dev.azure.com/fluidframework/internal/_workitems/edit/2118/

This PR allows running prettier in each sub-directory of release group
and packages. Instead of adding the script to the root directory,
developers can run `npm format` in any directory they are currently
working on which will fix entire files accordingly to the prettier
configuration.

There should be a separate PR to run prettier in `packages/dds`
directory.