Update deps on build-tools release group in client release group and independent packages (#12982)
Bumped dependencies in the following:

  client (release group)
  @fluidframework/common-definitions
  @fluidframework/common-utils
  @fluidframework/protocol-definitions
  @fluid-tools/api-markdown-documenter

Dependencies on these build-tools packages were updated:

  @fluid-tools/build-cli: ^0.6.0-109418
  @fluidframework/bundle-size-tools: ^0.6.0-109418
  @fluidframework/build-tools: ^0.6.0-109418

As part of the upgrade, the following additional changes were made:

- Use flub to generate typetests in client release group and independent
packages. Other release groups will be updated when they upgrade to the
new version of build-tools.
- type test baselines were set to the base **major** version for client
(2.0.0-internal.2.0.0), and the base **minor** version for independent
packages (varies). This was done to minimize changes. Once this is
merged I'll update the typetests to the appropriate versions on each
branch.
- Use flub for policy-check commands in r11s, build-tools, and azure.