Update bot config  (#13019)
This PR updates the bot config to not have check list:
`requireSpecificCheckRunsList` as the config doesn't seem to respond as
expected on the automation prs.

The bot was supposed to merge the PR when all the GitHub configured PR
checks are passed. There are times when we have Real service e2e and
stress test running on the automation pr which blocks the bot for
merging the PR in next.

In the [initial
commit](https://github.com/microsoft/FluidFramework/pull/12971),
`requireSpecificCheckRunsList` check was configured with the above
expectation. After the
[change](https://github.com/microsoft/FluidFramework/pull/12971) was
merged, the bot doesn't even seem to merge the PR in next.