Add threshold for non-runtime ops triggering summary heuristics (#13026)
We are summarizing too frequently when the summary only contains a few
non-system ops. There needs to be a threshold to control how many ops
since last summary are required before a non-system op can trigger a
summary.

https://portal.microsofticm.com/imp/v3/incidents/details/349754396/home

A threshold of 20 was chosen based on looking at telemetry. For every
summary with 100 ops or less, the average number of non-runtime ops is
below 20.