Fix get delta gap (#13035)
## Description

Users encounter GetDeltas_Exception {"error":"0x26d" } in the cases
where the logtail is missing the first seq number that is being
requested. Also, adding a check to verify that the seq numbers in the
logtail are contiguous.

If these conditions are not met, we want to fall back to get this data
from the server.