Prepare and generate typetests on main (#13029)
I manually updated the npm scripts to call flub with the correct
arguments etc. Relevant changes:

- Uses the `--pin` argument to pin the version in package.json to the
maximum matching released version. This should help with the lockfile
issues we have discovered.
- All generated files now include `.generated` in their name and old
files were deleted. This was detected as a rename by git.

See #12849 or `typetestDetails.md` for documentation on semantics of
these changes.

Then I did the following steps to configure the baselines and generate
tests.

1. `npm run typetests:prepare -- -b main --reset` - this reset all type
test overrides and set the baseline versions according to the main
branch config. For the client release group this is `~previousMinor`,
which corresponds to version 2.0.0-internal.2.1.0.
2. `npm i` - installed new versions of packages.
3. `npm run typetests:gen -- -b main` - this regenerated the typetests.
4. Ran a build and added back broken test overrides as needed.
5. Disabled tests in some packages like PropertyDDS. 

I also did roughly the same steps for the independent packages.