[GC-stress] Updated the test to send ops as per opRatePerMin config (#12401)
Note: This change is in a test branch not in main.

- Updated the tests to send ops as per the opRatePerMin config.
- The root data store does not send ops but just loads the child data
stores.
- Basically, there can be maximum of 8 data stores (excluding root) that
are referenced and sending ops. So, they each send opRatePerMin / 8 ops
per minute. The only disadvantage is that we may end up sending less
than opRatePerMin every minute since the actual number of data stores
can be < 8.
- This simplifies the calculation of how long the test will run. It will
also help us adjust the op rate as per the throttling limits.