Add test for legacy chunking (#12999)
Related to ADO:2465

There is a plan to refactor some of the op processing code inside the
container runtime, along with the upcoming plan to use chunking to
bypass the 1MB payload size limit for compressed batches. This implies
few changes in the way we process incoming ops at the runtime layer and
a resurrection of the code which produces chunked ops.

But before that happens, we need to ensure that the new work is not
breaking backwards compatibility wrt to processing legacy chunked ops,
as these might already be serialized in older document snapshots
(Excluding instances of really old clients collaborating on the same doc
with new clients). This test will make sure that we don't break the
basic scenario of processing chunked ops.

Also adding a bit of test infra to allow for installing specific older
versions of fluid.