Implement @fluid-internal/attributor (#12881)
## Description

This implements a first cut of Attributor functionality and the core
APIs. A couple design choice comments:

- I chose to decouple serialization of the attributor from the core
interface as it tends to make things more testable. Also makes it a fair
bit easier to play around with different serialization strategies before
we settle on one that we commit to back-compat for.
- I split the portion of attributor which associates information from
the portion that listens to the op stream and automatically adds two
attribution keys. This looks a little silly right now (it's a trivial
wrapper around `Map`), but my choice there has a couple more subtle
reasons:
- Once we add GC APIs, they should be implementable on the base
attributor. This will make it no longer completely trivial
- Having an Attributor implementation which works without an op stream
is convenient for testing (requires less mocking)
- We may eventually want attributors that don't automatically listen to
the op stream. The main use case I'm keeping in mind here is framework
support for attribution of external content (e.g. external copy-paste):
the attribution for a piece of content may conceptually contain users
that have never opened the document.

The main notable missing feature is GC. This PR doesn't make an attempt
to commit to the final serialization format.