Cherry picked buffer dependency fix from main to LTS branch [main->Lts] (#13095) (#13108)
# Work Item

## Description

This bug fix is to polyfill the ```buffer``` package. This bug causes
compiling any project that uses these libraries to fail. Upgrading to
webpack 5 removed the automatic NodeJS polyfills. I have added buffer as
a dependency to all the non-server packages that depend on safe buffer.

## Exception or Error
```
Compiled with problems:X

ERROR in ./node_modules/safe-buffer/index.js 2:13-30

Module not found: Error: Can't resolve 'buffer' in '<workspace>/node_modules\safe-buffer'

BREAKING CHANGE: webpack < 5 used to include polyfills for node.js core modules by default.
This is no longer the case. Verify if you need this module and configure a polyfill for it.

If you want to include a polyfill, you need to:
	- add a fallback 'resolve.fallback: { "buffer": require.resolve("buffer/") }'
	- install 'buffer'
If you don't want to include a polyfill, you can use an empty module like this:
	resolve.fallback: { "buffer": false }
```

## Steps to Reproduce Bug and Validate Solution

Follow any documentation to get started with fluid-framework. Preferably
with React or Angular but it should be the same for all JavaScript
client side projects

## Testing

We were able to import the packages ```fluid-framework``` and
```@fluidframework/tinylicious-client``` with my local changes into a
new React project we created in ```FluidFramework/examples/apps/```,
build and test successfully. We were unable to reproduce the errors
since they were always compiling in this directory. I will validate the
pre-release package for testing.

## Other information or known dependencies

[AB#1230](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1230)

[AB#1219](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1219)

[AB#1917](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/1917)
https://github.com/microsoft/FluidFramework/pull/11271