Tree: improve and fix cursor tests and fix a couple cursor bugs (#13027)
## Description

Cursor utils mishandled undefined and null node values. That is fixed in
this PR.

JsonCursor was difficult to construct from generic datasources (ex:
could not be constructed from JsonCompatible objects) due to issues with
Jsonable (causing infinite type recursion errors, or requiring casting
to `any`). Use of Jsonable has been removed to enable code like the
tests (and likely many real uses with work with generic json data
instead of specific types which happen to work with Jsonable).

JsonCursor has been fixed to correctly handle field access on null, and
a bug in its tests for this (which used the list of tests as the test
data instead of the contents of that list) has also been fixed.

Additionally the generic cursor test suite has been refactored and
improved so JsonCursor can use it to cover these cases and more (this
redundantly detected the same bugs).

This change was started to as a way to enable cursors which do not
support all possible trees to use more of the generic cursor test suite,
covering things like correct handling of problematic keys, working
traversal etc. Json cursor benefited from this since it is such a
cursor.

A follow up change should be made to prune the now redundant tests from
json cursor, but I think leaving those tests in to help show the bugs in
the old test suite (via the fixes in this change) is useful for helping
clarify what is going on in this PR and for adding confidence to the
changes to JsonCursor not just being due to subjective choices in the
new test suite but due to actual bugs in the old one.

This also moves the forest cursor tests into the forest test suite,
instead of in the object forest specific one.