ci: Cache pnpm when used (#13086)
This change updates the CI pipeline to cache pnpm's global store on a
per-release-group basis. For small release groups like build-tools , the
cache is of limited benefit, but I expect that the client release group
would benefit quite a bit (if we use pnpm there as well). This change is
done in a way that other release groups can switch to it if/when they
switch to pnpm.