Reset common-utils to 0.32.1 version in new release branch (#13135)
We want to release a patch of common-utils, 0.32.2. However, the commit
that common-utils 0.32.1 was release from is very old and doesn't build.
There have also been changes to common-utils on the lts branch in the
meantime.

To enable us to create a release branch for common-utils 0.32.x, this PR
resets the common-utils code to what it was at the 0.32.1 commit. That
is, I ran this command from the root of the repo on the lts branch:

```shell
git reset common-utils_v0.32.1 -- common/lib/common-utils/
git commit -m "reset common-utils"
git reset --hard
```

I then ran `policy-check:fix`, which made a small irrelevant change to
the package.json.

Once this PR is merged, the release/common-utils/0.32 branch should have
the correct common-utils code and also be releasable because the
pipeline will work.