Remove Set Package Version step (#13130)
## Description

Remove the `Set Package Version` step from Test Stability pipeline, it
is unnecessary for the testing stability purpose. Due to the recent
changes in `include-set-package-version.yml`, the `$(release)` can not
be none by default anymore. Please check
[AB2700](https://dev.azure.com/fluidframework/internal/_workitems/edit/2700/)
for more details

## Breaking Changes

N/A

## Testing

The updated yaml works well in test branch:
https://dev.azure.com/fluidframework/internal/_build/results?buildId=112310&view=results