Reset server to 0.1036.5000 version in new release branch (#13138)
We want to release a patch of server, 0.1036.5001. However, the commit
that server 0.1036.5001 was release from is very old and doesn't build.
There have also been changes to server on the lts branch in the
meantime.

To enable us to create a release branch for server 0.1036.5001, this PR
resets the common-utils code to what it was at the 0.1036.5001 commit.
That is, I ran this command from the root of the repo on the lts branch:

```shell
git reset server_v0.1036.5000 -- server/routerlicious/
git commit -m "reset server"
git reset --hard
```

Once this PR is merged, the release/server/0.1036.5000 branch should
have the correct server code and also be releasable because the pipeline
will work.