Tree: cleanup cursor tests, and fix json cursor bug (#13122)
## Description

This deduplicates some of the json cursor specific tests with the ones
in the cursor test suite.

As part of this, testing of keys on specific types of nodes was reworked
so it did not depend on builders from jsonableTrees so that they could
be applied to specialized cursors like JsonCursor. This rework checked
for keys on all the test tree nodes, and this exposed a bug which the
old test suite missed: keys like "0" were showing up on strings in the
json domain when using named field access but not in field iteration. A
fix for this bug is also included.

Co-authored-by: Jenn <jennle@microsoft.com>