fix(run:bundleStats): Take path to dangerfile instead of directory
The dangerfile path had some "magic strings" in the code, so this change makes the full path to the
dangerfile the argument instead of a folder containing a dangerfile at a specific place inside it.
The default dangerfile is still used in all known uses, so the real meaningful value is the default.

BREAKING CHANGE: The --dirname argument has been removed. There is now a --dangerfile argument that
defaults to the built-in dangerfile but can be customized if needed.