[TaskManager] Cleanup TODOs (#13169)
## Description

This PR cleans up miscellaneous TODO items for TaskManager, such as the
removal of debugging events/functions.

Note: This PR cherry-picks from
https://github.com/microsoft/FluidFramework/pull/13014, but with the
notable re-inclusion of `this.scrubClientsNotInQuorum()` and filtering
`this.taskQueues` during summarization. The removal of these proved to
be overly optimistic since we saw regressions in stress tests after the
previously mentioned PR was merged. We may create a follow-up task in
the future to further investigate the root cause.

## Breaking Changes

This PR removes `TaskManager._getTaskQueues()`. Although this is
technically a breaking change, this function was always intended to be
used for debugging purposes. Additionally, TaskManager is currently an
experimental package.

## Reviewer Guidance


[AB#2422](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2422)

[AB#2752](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/2752)