feat: Promote `import/no-nodejs-modules` to an error (#13241)
Rule documentation:
https://github.com/import-js/eslint-plugin-import/blob/main/docs/rules/no-nodejs-modules.md

This PR promotes the above rule from a `warning` to an `error` in our
base eslint configuration, but exempts imports from "events" due to the
sheer number of uses in the codebase.

This PR also "pre-fixes" violations across the codebase by disabling the
rule in packages / inline as needed without making any code changes.
Code changes may be made in the future as appropriate to remove rule
overrides. But for now, the primary goal is to prevent future
violations.

### Breaking Change

This PR promotes an existing `warn`-level rule to an `error`.