Tree: expose `getPrimaryField` and refactor EditableTree implementation & tests (#13271)
This PR exposes `getPrimaryField` and refactors the EditableTree and its
tests.

The main cause of this is to prepare the code base for the upcoming
implementation of "simple assignments" and hence to make the latter
smaller i.e. better for a review.

Some points worth mentioning:

#### API changes
- `getPrimaryField` is exposed to let users handle corresponding cases
(e.g. working with arrays)
- added schema to `EditableTreeContext` to shorten the
`context.forest.schema` way of accessing it

#### Implementation refactoring
- `getFieldSchema` is wrapped within the EditableTree proxy - this has
no use currently, but will be needed in "simple assignments"

#### Test refactoring

Changes in tests are mainly caused by improving the mock data. Those
improvements are the result of the "Property Inspector" PoC demoed for
various user groups, introducing among others the type `Bool` and
schemas (and types) for "array of array" usecases.