update minimumSequenceNumber (#13062)
## Description
Task
[1811](https://dev.azure.com/fluidframework/internal/_sprints/taskboard/Team%20Taylor/internal/CY22Q4/2022.11/Team%20Taylor%20-%202022.11?workitem=1811)

The EditManager class maintains a list of fully rebased sequenced edits
to represent the master/trunk branch of the live collaboration session.
This list only needs to contain these edits as far back as the Fluid
collaboration window. Right now, the list is never trimmed to remove
edits that lie outside of the collaboration window, which leads to
unbounded memory growth as the collab session stretches over time.

The goal of this task is to remedy the unbounded memory growth by
removing from the trunk those edits/commits that lie outside the
collaboration window. To do this, the edit manager will need to be told
about the progression of the collaboration window. This information is
exposed as ISequencedDocumentMessage.minimumSequenceNumber which is
passed to processCore (in sharedTreeCore.ts).

## Breaking Changes

No

## Reviewer Guidance

### Questions
1. Currently, all this PR does is remove the first n elements in the
`trunk` field of the `editManager` class, with sequence numbers less
than the minimumSequenceNumber. Are there any other fields that should
also be updated?

Co-authored-by: Yann Achard <achardy@microsoft.com>